Get Speed into fragment with the same Viewmodel: get the current opened websocket

+++ app/src/main/java/com/seigneur/gauvain/needforstream/data/model/Car.kt
@@ -12,5 +12,5 @@ data class Car(
     @SerializedName("Cv")
     var Cv: Int,
     @SerializedName("CurrentSpeed")
-    var CurrentSpeed: Int)
+    var CurrentSpeed: Float)
